Quite a lot of changes - mainly fixes.
One breaking change:-
Paging no longer requires the data function to return a count of the total number of records, but it does now need
the pageSize setting to be given a value.

Combo mode - when wrapper gets the focus, set focus instead to the combo text box, so user can type in it straight away.

Allow enter and escape key events to propagate when the pop-up isn't visible.

Fix to width when no items are found.

In loadData, fixed an issue where items could sometimes appear twice.

Fire the ac-change event when a selection is cleared.

Fixed an error that was occurring when there was no selected item.

CSS for dropdown gif - added no-repeat.

Fix for 'cannot set length property of undefined'

Added support for watermark text (when in combo mode) via the "placeholderText" setting.

Added new pageSize setting. Removed need for data function to return a count of the total number of records to enable paging. Now shows "Load more" message when fewer than [pageSize] items are returned by the host data load function.

Efficiency. No longer calls the data load function if the search text contains the previous search text value, because the necessary data is already present and can be filtered client-side.

Watch for changes on the model. If set to blank/null values, clear the selection.

Data wasn't always loading when previousSearchText was an empty string

Take placeholderText into account when setting dropdown width.

Fixed a bug in with IE where an undefined check was needed.

Avoid error if a data item is null, or has a null text value.

When the control wrapper gets the focus, if in combo mode, transfer the focus to the combo textbox and select the text in it.

Clear any filtering when the dropdown is opened.

Select item if the search text exactly matches one item, even if there are others that it partly matches.

Eliminated horizontal scrollbar that was sometimes showing.

Correction to filtering, so it acts as expected when search text characters are deleted.

Removed spurious ng-model attribute from li elements. This was causing angular to set up a $watch for every item in every instance, which caused performance problems on pages with a lot of selects.

Fixed issue with drop downs reporting "No records found".

Fixed problem with filter text reverting while typing, if the new value still matched a single item.

Fix for error when pressing delete key, due to setting event.stopPropagation to true, instead of calling the event.stopPropagation function.

When pasting or typing a value that exactly matches an item in the dropdown, ensure that item is selected.

Removed ng-keydown for the combo textbox, which was not needed here as well as on the container div and caused the keyHandler to fire twice.

Fixed to not fire the ac-change event when setting an initial selection.

Correctly fire the ac-change event when a value is first set.

Don't set focus when the pop-up is not shown. When the use selects an item and the dropdown closes, focus is given to the acute-select container,
which in turn gives focus to the combo textbox. This was, however, also being done when the selection was set programmatically.

ng-change event wasn't always firing. Now always fires if the pop-up list is open, but not when it's closed, i.e. when initialising the selection.