Bug fix release + 2 new settings and 1 small breaking change (pageSize)
Breaking change - Paging no longer requires the data function to return a count of the total number of records, but it does now need the pageSize setting to be given a value.

Show message "Press Enter to Add." when in combo mode and no items are found, but AllowCustomText is true.

Fix to acSelectOnFocus directive. Select text on focus unless ac-select-on-focus is false (boolean) or 'false' (string).

Filtering didn't update/clear when deleting last character

Clear items before loading new ones UNLESS paging is enabled and an offset was passed to data function. N.B. this means that script using ac-select must return the offset value to the callback function for paging to work.

Added 'ac-focus-when' attribute. When expression evaluates as *true* focus is given to either the combo or search text box.

Was sometimes not filtering correctly when typing quickly and loading data asynchronously.

Avoid intermittent error in ac-focus directive, by using safeApply rather than $scope.$apply.

Fix issue where text was re-appearing after keying a backspace.

Added a maxWidth setting. Truncate text if it exceeds this maximum width, but show the full text as a tooltip.

Fix - wasn't working with a simple array of string values.